Add files via upload
This was the 1st project I collaborated with a colleague remotely over 3 time zones away.  I was brought onboard to his idea of using SQL Server to create an even number of high-selling and low-selling dealers (400+ total) for each sales rep (~16) nation-wide.  The goal in the end was to use the geography attribute to help calculate distance for each.
After 4 - 6 months the project was a complete failure.  Because of that, it's still one of my favorite projects.